requirements.txt
pytorch==2.7.1
numpy
pyrtklib
pymap3d
pandas
tqdm
matplotlib
pyrtklib